建设工单保存功能

+++ src/main/java/com/dgpalife/resourcemanagement/service/OrderService.java
@@ -3,5 +3,5 @@ package com.dgpalife.resourcemanagement.service;
 import com.dgpalife.resourcemanagement.model.Order;
 
 public interface OrderService {
-    Long createOrder(Order order);
+    Integer createOrder(Order order);
 }
